+++ server/models/index.js
@@ -1,6 +1,6 @@
 const Neode = require('neode');
-const Person = require('./person/person.model');
-const Certification = require('./certification/certification.model');
+const Person = require('./person.model');
+const Certification = require('./certification.model');
 
 const neode = new Neode.fromEnv().with({
   Person: Person,
